Separate JITServerHelpers from JITServerCompilationThread
[skip ci]
Extract `JITServer` helper and utility functions out of
`JITServerCompilationThread`.
