+++ runtime/compiler/control/JITServerCompilationThread.cpp
@@ -21,64 +21,21 @@
  *******************************************************************************/
 
 #include "control/JITServerCompilationThread.hpp"
-#include "vmaccess.h"
-#include "runtime/J9VMAccess.hpp"
-#include "env/VMAccessCriticalSection.hpp"
-#include "control/CompilationRuntime.hpp"
-#include "compile/Compilation.hpp"
-#include "control/MethodToBeCompiled.hpp"
+
 #include "codegen/CodeGenerator.hpp"
-#include "runtime/CodeCache.hpp"
+#include "control/CompilationRuntime.hpp"
+#include "control/MethodToBeCompiled.hpp"
+#include "env/ClassTableCriticalSection.hpp"
+#include "env/VMAccessCriticalSection.hpp"
 #include "env/JITServerPersistentCHTable.hpp"
-#include "env/JITServerCHTable.hpp"
-#include "env/ClassTableCriticalSection.hpp"   // for ClassTableCriticalSection
-#include "exceptions/RuntimeFailure.hpp"       // for CHTableCommitFailure
-#include "runtime/IProfiler.hpp"               // for TR_IProfiler
-#include "runtime/JITServerIProfiler.hpp"      // for TR_ContiguousIPMethodHashTableEntry
-#include "j9port.h"                            // for j9time_current_time_millis
-#include "env/j9methodServer.hpp"
-#include "exceptions/AOTFailure.hpp"           // for AOTFailure
-#include "env/TRMemory.hpp"                    // for jitPersistentAlloc and jitPersistentFree
-#include "runtime/CodeCacheManager.hpp"
+#include "runtime/CodeCache.hpp"
 #include "runtime/CodeCacheExceptions.hpp"
+#include "runtime/J9VMAccess.hpp"
 #include "runtime/RelocationTarget.hpp"
 #include "jitprotos.h"
-
-#if defined(J9VM_OPT_SHARED_CLASSES)
-#include "j9jitnls.h"
-#endif
+#include "vmaccess.h"
 
 extern TR::Monitor *assumptionTableMutex;
-
-uint32_t serverMsgTypeCount[JITServer::MessageType_ARRAYSIZE] = {};
-
-uint64_t JITServerHelpers::_waitTime = 1000;
-bool JITServerHelpers::_serverAvailable = true;
-uint64_t JITServerHelpers::_nextConnectionRetryTime = 0;
-TR::Monitor * JITServerHelpers::_clientStreamMonitor = NULL;
-
-
-// insertIntoOOSequenceEntryList needs to be executed with sequencingMonitor in hand
-// This method belongs to ClientSessionData, but is temporarily moved here to be able
-// to push the ClientSessionData related code as a standalone piece
-void 
-JITServerHelpers::insertIntoOOSequenceEntryList(ClientSessionData *clientData, TR_MethodToBeCompiled *entry)
-   {
-   uint32_t seqNo = ((TR::CompilationInfoPerThreadRemote*)(entry->_compInfoPT))->getSeqNo();
-   TR_MethodToBeCompiled *crtEntry = clientData->getOOSequenceEntryList();
-   TR_MethodToBeCompiled *prevEntry = NULL;
-   while (crtEntry && seqNo > ((TR::CompilationInfoPerThreadRemote*)(crtEntry->_compInfoPT))->getSeqNo())
-      {
-      prevEntry = crtEntry;
-      crtEntry = crtEntry->_next;
-      }
-   entry->_next = crtEntry;
-   if (prevEntry)
-      prevEntry->_next = entry;
-   else
-      clientData->setOOSequenceEntryList(entry);
-   }
-
 
 // TODO: this method is copied from runtime/jit_vm/ctsupport.c,
 // in the future it's probably better to make that method publicly accessible
@@ -86,7 +43,7 @@ static UDATA
 findField(J9VMThread *vmStruct, J9ConstantPool *constantPool, UDATA index, BOOLEAN isStatic, J9Class **declaringClass)
    {
    J9JavaVM *javaVM = vmStruct->javaVM;
-   J9ROMFieldRef *romRef; 
+   J9ROMFieldRef *romRef;
    J9ROMClassRef *classRef; /* ROM class of the field */
    U_32 classRefCPIndex;
    J9UTF8 *classNameUTF;
@@ -136,68 +93,9 @@ findField(J9VMThread *vmStruct, J9ConstantPool *constantPool, UDATA index, BOOLE
    return result;
    }
 
-size_t methodStringsLength(J9ROMMethod *method)
-   {
-   J9UTF8 *name = J9ROMMETHOD_NAME(method);
-   J9UTF8 *sig = J9ROMMETHOD_SIGNATURE(method);
-   return name->length + sig->length + 2 * sizeof(U_16);
-   }
 
-// Packs a ROMClass into a std::string to be transfered to the server.
-// The name and signature of all methods are appended to the end of the cloned class body and the
-// self referential pointers to them are updated to deal with possible interning. The method names
-// and signature are needed on the server but may be interned globally on the client.
-std::string packROMClass(J9ROMClass *origRomClass, TR_Memory *trMemory)
-   {
-   //JITServer TODO: Add comments
-   J9UTF8 *className = J9ROMCLASS_CLASSNAME(origRomClass);
-   size_t classNameSize = className->length + sizeof(U_16);
-
-   J9ROMMethod *romMethod = J9ROMCLASS_ROMMETHODS(origRomClass);
-   size_t totalSize = origRomClass->romSize + classNameSize;
-   for (size_t i = 0; i < origRomClass->romMethodCount; ++i)
-      {
-      totalSize += methodStringsLength(romMethod);
-      romMethod = nextROMMethod(romMethod);
-      }
-
-   J9ROMClass *romClass = (J9ROMClass *)trMemory->allocateHeapMemory(totalSize);
-   if (!romClass)
-      throw std::bad_alloc();
-   memcpy(romClass, origRomClass, origRomClass->romSize);
-
-   uint8_t *curPos = ((uint8_t *)romClass) + romClass->romSize;
-
-   memcpy(curPos, className, classNameSize);
-   NNSRP_SET(romClass->className, curPos);
-   curPos += classNameSize;
-
-   romMethod = J9ROMCLASS_ROMMETHODS(romClass);
-   J9ROMMethod *origRomMethod = J9ROMCLASS_ROMMETHODS(origRomClass);
-   for (size_t i = 0; i < romClass->romMethodCount; ++i)
-      {
-      J9UTF8 *field = J9ROMMETHOD_NAME(origRomMethod);
-      size_t fieldLen = field->length + sizeof(U_16);
-      memcpy(curPos, field, fieldLen);
-      NNSRP_SET(romMethod->nameAndSignature.name, curPos);
-      curPos += fieldLen;
-
-      field = J9ROMMETHOD_SIGNATURE(origRomMethod);
-      fieldLen = field->length + sizeof(U_16);
-      memcpy(curPos, field, fieldLen);
-      NNSRP_SET(romMethod->nameAndSignature.signature, curPos);
-      curPos += fieldLen;
-
-      romMethod = nextROMMethod(romMethod);
-      origRomMethod = nextROMMethod(origRomMethod);
-      }
-
-   std::string romClassStr((char *) romClass, totalSize);
-   trMemory->freeMemory(romClass, heapAlloc);
-   return romClassStr;
-   }
-
-void handler_IProfiler_profilingSample(JITServer::ClientStream *client, TR_J9VM *fe, TR::Compilation *comp)
+static void
+handler_IProfiler_profilingSample(JITServer::ClientStream *client, TR_J9VM *fe, TR::Compilation *comp)
    {
    auto recv = client->getRecvData<TR_OpaqueMethodBlock*, uint32_t, uintptrj_t>();
    auto method = std::get<0>(recv);
@@ -249,7 +147,8 @@ void handler_IProfiler_profilingSample(JITServer::ClientStream *client, TR_J9VM 
       }
    }
 
-static bool handleServerMessage(JITServer::ClientStream *client, TR_J9VM *fe, JITServer::MessageType &response)
+static bool
+handleServerMessage(JITServer::ClientStream *client, TR_J9VM *fe, JITServer::MessageType &response)
    {
    using JITServer::MessageType;
    TR::CompilationInfoPerThread *compInfoPT = fe->_compInfoPT;
@@ -271,7 +170,7 @@ static bool handleServerMessage(JITServer::ClientStream *client, TR_J9VM *fe, JI
    acquireVMAccessNoSuspend(vmThread);
 
    // Update statistics for server message type
-   serverMsgTypeCount[response] += 1;
+   JITServerHelpers::serverMsgTypeCount[response] += 1;
 
    // If JVM has unloaded classes inform the server to abort this compilation
    uint8_t interruptReason = compInfoPT->compilationShouldBeInterrupted();
@@ -454,7 +353,7 @@ static bool handleServerMessage(JITServer::ClientStream *client, TR_J9VM *fe, JI
                   compareFields = true;
                   }
                }
-            }      
+            }
          if (compareFields)
             {
              f1 = findField(fe->vmThread(), cp1, cpIndex1, isStatic, &declaringClass1);
@@ -1062,10 +961,10 @@ static bool handleServerMessage(JITServer::ClientStream *client, TR_J9VM *fe, JI
          TR_OpaqueClassBlock *clazz = std::get<0>(recv);
          std::string fieldName = std::get<1>(recv);
          std::string sigName = std::get<2>(recv);
-         client->write(response, fe->getStaticFieldAddress(clazz, 
-                  reinterpret_cast<unsigned char*>(const_cast<char*>(fieldName.c_str())), 
-                  fieldName.length(), 
-                  reinterpret_cast<unsigned char*>(const_cast<char*>(sigName.c_str())), 
+         client->write(response, fe->getStaticFieldAddress(clazz,
+                  reinterpret_cast<unsigned char*>(const_cast<char*>(fieldName.c_str())),
+                  fieldName.length(),
+                  reinterpret_cast<unsigned char*>(const_cast<char*>(sigName.c_str())),
                   sigName.length()));
          }
          break;
@@ -1229,7 +1128,7 @@ static bool handleServerMessage(JITServer::ClientStream *client, TR_J9VM *fe, JI
          TR_ResolvedJ9JITServerMethodInfo methodInfo;
          // if in AOT mode, create a relocatable method mirror
          TR_ResolvedJ9JITServerMethod::createResolvedMethodMirror(methodInfo, method, vTableSlot, owningMethod, fe, trMemory);
-         
+
          client->write(response, methodInfo);
          }
          break;
@@ -1320,8 +1219,8 @@ static bool handleServerMessage(JITServer::ClientStream *client, TR_J9VM *fe, JI
             TR::VMAccessCriticalSection resolveStaticMethodRef(fe);
             ramMethod = jitResolveStaticMethodRef(fe->vmThread(), method->cp(), cpIndex, J9_RESOLVE_FLAG_JIT_COMPILE_TIME);
             }
-        
-         // create a mirror right away
+
+            // create a mirror right away
          TR_ResolvedJ9JITServerMethodInfo methodInfo;
          if (ramMethod)
             TR_ResolvedJ9JITServerMethod::createResolvedMethodFromJ9MethodMirror(methodInfo, (TR_OpaqueMethodBlock *) ramMethod, 0, method, fe, trMemory);
@@ -1346,7 +1245,7 @@ static bool handleServerMessage(JITServer::ClientStream *client, TR_J9VM *fe, JI
             if (ramMethod)
                TR_ResolvedJ9JITServerMethod::createResolvedMethodFromJ9MethodMirror(methodInfo, (TR_OpaqueMethodBlock *) ramMethod, 0, method, fe, trMemory);
             }
-         
+
          client->write(response, ramMethod, methodInfo);
          }
          break;
@@ -1373,7 +1272,7 @@ static bool handleServerMessage(JITServer::ClientStream *client, TR_J9VM *fe, JI
          J9ConstantPool *cp = (J9ConstantPool*)literals;
          I_32 cpIndex = std::get<2>(recv);
          bool resolvedInCP = false;
-         
+
          // only call the resolve if unresolved
          J9Method * ramMethod = 0;
          UDATA vTableIndex = (((J9RAMVirtualMethodRef*) literals)[cpIndex]).methodIndexAndArgCount;
@@ -1400,10 +1299,10 @@ static bool handleServerMessage(JITServer::ClientStream *client, TR_J9VM *fe, JI
          if (vTableIndex)
             {
             TR_OpaqueMethodBlock *method = (TR_OpaqueMethodBlock *) ramMethod;
-           
+
             TR_ResolvedJ9JITServerMethodInfo methodInfo;
             TR_ResolvedJ9JITServerMethod::createResolvedMethodFromJ9MethodMirror(methodInfo, (TR_OpaqueMethodBlock *) ramMethod, (uint32_t) vTableIndex, owningMethod, fe, trMemory);
-                        
+
             client->write(response, ramMethod, vTableIndex, resolvedInCP, methodInfo);
             }
          else
@@ -2595,7 +2494,7 @@ remoteCompile(
    JITServer::ClientStream *client = enableJITServerPerCompConn ? NULL : compInfoPT->getClientStream();
    if (!client)
       {
-      try 
+      try
          {
          if (JITServerHelpers::isServerAvailable())
             {
@@ -2648,7 +2547,6 @@ remoteCompile(
    std::string optionsStr = TR::Options::packOptions(compiler->getOptions());
    std::string recompMethodInfoStr = compiler->isRecompilationEnabled() ? std::string((char *) compiler->getRecompilationInfo()->getMethodInfo(), sizeof(TR_PersistentMethodInfo)) : std::string();
 
-   
    compInfo->getSequencingMonitor()->enter();
    // Collect the list of unloaded classes
    std::vector<TR_OpaqueClassBlock*> unloadedClasses(compInfo->getUnloadedClassesTempList()->begin(), compInfo->getUnloadedClassesTempList()->end());
@@ -2673,7 +2571,7 @@ remoteCompile(
    try
       {
       // release VM access just before sending the compilation request
-      // message just in case we block in the write operation   
+      // message just in case we block in the write operation
       releaseVMAccess(vmThread);
 
       if (TR::Options::getVerboseOption(TR_VerboseJITServer))
@@ -2886,7 +2784,7 @@ remoteCompile(
    return metaData;
    }
 
-TR_MethodMetaData * 
+TR_MethodMetaData *
 remoteCompilationEnd(
    J9VMThread * vmThread,
    TR::Compilation *comp,
@@ -2994,7 +2892,7 @@ remoteCompilationEnd(
             returnCode = compilationAotRelocationInterrupted;
             }
 
-         if (relocatedMetaData) 
+         if (relocatedMetaData)
             {
             if (TR::Options::getVerboseOption(TR_VerboseJITServer))
                {
@@ -3029,7 +2927,7 @@ remoteCompilationEnd(
       else
          {
          // AOT compilations can fail on purpose because we want to load
-         // the AOT body later on. This case is signalled by having a successful compilation 
+         // the AOT body later on. This case is signalled by having a successful compilation
          // but canRelocateMethod == false
          // We still need metadata, because metaData->startPC != 0 indicates that compilation
          // didn't actually fail.
@@ -3114,440 +3012,6 @@ outOfProcessCompilationEnd(
       }
    }
 
-void printJITServerMsgStats(J9JITConfig *jitConfig)
-   {
-   PORT_ACCESS_FROM_JITCONFIG(jitConfig);
-   j9tty_printf(PORTLIB, "JITServer Message Type Statistics:\n");
-   j9tty_printf(PORTLIB, "Type# #called TypeName\n");
-   const ::google::protobuf::EnumDescriptor *descriptor = JITServer::MessageType_descriptor();
-   for (int i = 0; i < JITServer::MessageType_ARRAYSIZE; ++i)
-      {
-      if (serverMsgTypeCount[i] > 0)
-         j9tty_printf(PORTLIB, "#%04d %7u %s\n", i, serverMsgTypeCount[i], descriptor->FindValueByNumber(i)->name().c_str());
-      }
-   }
-
-void printJITServerCHTableStats(J9JITConfig *jitConfig, TR::CompilationInfo *compInfo)
-   {
-#ifdef COLLECT_CHTABLE_STATS
-   PORT_ACCESS_FROM_JITCONFIG(jitConfig);
-   j9tty_printf(PORTLIB, "JITServer CHTable Statistics:\n");
-   if (compInfo->getPersistentInfo()->getRemoteCompilationMode() == JITServer::CLIENT)
-      {
-      JITClientPersistentCHTable *table = (JITClientPersistentCHTable*)compInfo->getPersistentInfo()->getPersistentCHTable();
-      j9tty_printf(PORTLIB, "Num updates sent: %d (1 per compilation)\n", table->_numUpdates);
-      if (table->_numUpdates)
-         {
-         j9tty_printf(PORTLIB, "Num commit failures: %d. Average per compilation: %f\n", table->_numCommitFailures, table->_numCommitFailures / float(table->_numUpdates));
-         j9tty_printf(PORTLIB, "Num classes updated: %d. Average per compilation: %f\n", table->_numClassesUpdated, table->_numClassesUpdated / float(table->_numUpdates));
-         j9tty_printf(PORTLIB, "Num classes removed: %d. Average per compilation: %f\n", table->_numClassesRemoved, table->_numClassesRemoved / float(table->_numUpdates));
-         j9tty_printf(PORTLIB, "Total update bytes: %d. Compilation max: %d. Average per compilation: %f\n", table->_updateBytes, table->_maxUpdateBytes, table->_updateBytes / float(table->_numUpdates));
-         }
-      }
-   else if (compInfo->getPersistentInfo()->getRemoteCompilationMode() == JITServer::SERVER)
-      {
-      JITServerPersistentCHTable *table = (JITServerPersistentCHTable*)compInfo->getPersistentInfo()->getPersistentCHTable();
-      j9tty_printf(PORTLIB, "Num updates received: %d (1 per compilation)\n", table->_numUpdates);
-      if (table->_numUpdates)
-         {
-         j9tty_printf(PORTLIB, "Num classes updated: %d. Average per compilation: %f\n", table->_numClassesUpdated, table->_numClassesUpdated / float(table->_numUpdates));
-         j9tty_printf(PORTLIB, "Num classes removed: %d. Average per compilation: %f\n", table->_numClassesRemoved, table->_numClassesRemoved / float(table->_numUpdates));
-         j9tty_printf(PORTLIB, "Num class info queries: %d. Average per compilation: %f\n", table->_numQueries, table->_numQueries / float(table->_numUpdates));
-         j9tty_printf(PORTLIB, "Total update bytes: %d. Compilation max: %d. Average per compilation: %f\n", table->_updateBytes, table->_maxUpdateBytes, table->_updateBytes / float(table->_numUpdates));
-         }
-      }
-#endif
-   }
-
-void printJITServerCacheStats(J9JITConfig *jitConfig, TR::CompilationInfo *compInfo)
-   {
-   PORT_ACCESS_FROM_JITCONFIG(jitConfig);
-   if (compInfo->getPersistentInfo()->getRemoteCompilationMode() == JITServer::SERVER)
-      {
-      auto clientSessionHT = compInfo->getClientSessionHT();
-      clientSessionHT->printStats();
-      }
-   }
-
-
-void 
-JITServerHelpers::cacheRemoteROMClass(ClientSessionData *clientSessionData, J9Class *clazz, J9ROMClass *romClass, ClassInfoTuple *classInfoTuple)
-   {
-   ClientSessionData::ClassInfo classInfo;
-   OMR::CriticalSection cacheRemoteROMClass(clientSessionData->getROMMapMonitor());
-   auto it = clientSessionData->getROMClassMap().find((J9Class*)clazz);
-   if (it == clientSessionData->getROMClassMap().end())
-      {
-      JITServerHelpers::cacheRemoteROMClass(clientSessionData, clazz, romClass, classInfoTuple, classInfo);
-      }
-   }
-
-void
-JITServerHelpers::cacheRemoteROMClass(ClientSessionData *clientSessionData, J9Class *clazz, J9ROMClass *romClass, ClassInfoTuple *classInfoTuple, ClientSessionData::ClassInfo &classInfoStruct)
-   {
-   ClassInfoTuple &classInfo = *classInfoTuple;
-
-   classInfoStruct._romClass = romClass;
-   classInfoStruct._methodsOfClass = std::get<1>(classInfo);
-   J9Method *methods = classInfoStruct._methodsOfClass;
-   classInfoStruct._baseComponentClass = std::get<2>(classInfo);
-   classInfoStruct._numDimensions = std::get<3>(classInfo);
-   classInfoStruct._remoteROMStringsCache = NULL;
-   classInfoStruct._fieldOrStaticNameCache = NULL;
-   classInfoStruct._parentClass = std::get<4>(classInfo);
-   auto &tmpInterfaces = std::get<5>(classInfo);
-   classInfoStruct._interfaces = new (PERSISTENT_NEW) PersistentVector<TR_OpaqueClassBlock *>
-      (tmpInterfaces.begin(), tmpInterfaces.end(),
-       PersistentVector<TR_OpaqueClassBlock *>::allocator_type(TR::Compiler->persistentAllocator()));
-   auto &methodTracingInfo = std::get<6>(classInfo);
-   classInfoStruct._classHasFinalFields = std::get<7>(classInfo);
-   classInfoStruct._classDepthAndFlags = std::get<8>(classInfo);
-   classInfoStruct._classInitialized = std::get<9>(classInfo);
-   classInfoStruct._byteOffsetToLockword = std::get<10>(classInfo);
-   classInfoStruct._leafComponentClass = std::get<11>(classInfo);
-   classInfoStruct._classLoader = std::get<12>(classInfo);
-   classInfoStruct._hostClass = std::get<13>(classInfo);
-   classInfoStruct._componentClass = std::get<14>(classInfo);
-   classInfoStruct._arrayClass = std::get<15>(classInfo);
-   classInfoStruct._totalInstanceSize = std::get<16>(classInfo);
-   classInfoStruct._classOfStaticCache = NULL;
-   classInfoStruct._constantClassPoolCache = NULL;
-   classInfoStruct._remoteRomClass = std::get<17>(classInfo);
-   classInfoStruct._fieldAttributesCache = NULL;
-   classInfoStruct._staticAttributesCache = NULL;
-   classInfoStruct._fieldAttributesCacheAOT = NULL;
-   classInfoStruct._staticAttributesCacheAOT = NULL;
-   classInfoStruct._constantPool = (J9ConstantPool *)std::get<18>(classInfo);
-   classInfoStruct._jitFieldsCache = NULL;
-   classInfoStruct._classFlags = std::get<19>(classInfo);
-   classInfoStruct._fieldOrStaticDeclaringClassCache = NULL;
-   classInfoStruct._J9MethodNameCache = NULL;
-   clientSessionData->getROMClassMap().insert({ clazz, classInfoStruct});
-
-   uint32_t numMethods = romClass->romMethodCount;
-   J9ROMMethod *romMethod = J9ROMCLASS_ROMMETHODS(romClass);
-   for (uint32_t i = 0; i < numMethods; i++)
-      {
-      clientSessionData->getJ9MethodMap().insert({&methods[i], 
-            {romMethod, NULL, static_cast<bool>(methodTracingInfo[i]), (TR_OpaqueClassBlock *)clazz, false}});
-      romMethod = nextROMMethod(romMethod);
-      }
-   }
-
-J9ROMClass *
-JITServerHelpers::getRemoteROMClassIfCached(ClientSessionData *clientSessionData, J9Class *clazz)
-   {
-   OMR::CriticalSection getRemoteROMClass(clientSessionData->getROMMapMonitor());
-   auto it = clientSessionData->getROMClassMap().find(clazz);
-   return (it == clientSessionData->getROMClassMap().end()) ? NULL : it->second._romClass;
-   }
-
-
-
-JITServerHelpers::ClassInfoTuple
-JITServerHelpers::packRemoteROMClassInfo(J9Class *clazz, J9VMThread *vmThread, TR_Memory *trMemory)
-   {
-   // Always use the base VM here.
-   // If this method is called inside AOT compilation, TR_J9SharedCacheVM will
-   // attempt validation and return NULL for many methods invoked here.
-   // We do not want that, because these values will be cached and later used in non-AOT
-   // compilations, where we always need a non-NULL result.
-   TR_J9VM *fe = (TR_J9VM *) TR_J9VMBase::get(vmThread->javaVM->jitConfig, vmThread);
-   J9Method *methodsOfClass = (J9Method*) fe->getMethods((TR_OpaqueClassBlock*) clazz);
-   int32_t numDims = 0;
-   TR_OpaqueClassBlock *baseClass = fe->getBaseComponentClass((TR_OpaqueClassBlock *) clazz, numDims);
-   TR_OpaqueClassBlock *parentClass = fe->getSuperClass((TR_OpaqueClassBlock *) clazz);
-
-   uint32_t numMethods = clazz->romClass->romMethodCount;
-   std::vector<uint8_t> methodTracingInfo;
-   methodTracingInfo.reserve(numMethods);
-   for(uint32_t i = 0; i < numMethods; ++i)
-      {
-      methodTracingInfo.push_back(static_cast<uint8_t>(fe->isMethodTracingEnabled((TR_OpaqueMethodBlock *) &methodsOfClass[i])));
-      }
-   bool classHasFinalFields = fe->hasFinalFieldsInClass((TR_OpaqueClassBlock *)clazz);
-   uintptrj_t classDepthAndFlags = fe->getClassDepthAndFlagsValue((TR_OpaqueClassBlock *)clazz);
-   bool classInitialized =  fe->isClassInitialized((TR_OpaqueClassBlock *)clazz);
-   uint32_t byteOffsetToLockword = fe->getByteOffsetToLockword((TR_OpaqueClassBlock *)clazz);
-   TR_OpaqueClassBlock * leafComponentClass = fe->getLeafComponentClassFromArrayClass((TR_OpaqueClassBlock *)clazz);
-   void * classLoader = fe->getClassLoader((TR_OpaqueClassBlock *)clazz);
-   TR_OpaqueClassBlock * hostClass = fe->convertClassPtrToClassOffset(clazz->hostClass);
-   TR_OpaqueClassBlock * componentClass = fe->getComponentClassFromArrayClass((TR_OpaqueClassBlock *)clazz);
-   TR_OpaqueClassBlock * arrayClass = fe->getArrayClassFromComponentClass((TR_OpaqueClassBlock *)clazz);
-   uintptrj_t totalInstanceSize = clazz->totalInstanceSize;
-   uintptrj_t cp = fe->getConstantPoolFromClass((TR_OpaqueClassBlock *)clazz);
-   uintptrj_t classFlags = fe->getClassFlagsValue((TR_OpaqueClassBlock *)clazz);
-   return std::make_tuple(packROMClass(clazz->romClass, trMemory), methodsOfClass, baseClass, numDims, parentClass, TR::Compiler->cls.getITable((TR_OpaqueClassBlock *) clazz), methodTracingInfo, classHasFinalFields, classDepthAndFlags, classInitialized, byteOffsetToLockword, leafComponentClass, classLoader, hostClass, componentClass, arrayClass, totalInstanceSize, clazz->romClass, cp, classFlags);
-   }
-
-J9ROMClass *
-JITServerHelpers::romClassFromString(const std::string &romClassStr, TR_PersistentMemory *trMemory)
-   {
-   auto romClass = (J9ROMClass *)(trMemory->allocatePersistentMemory(romClassStr.size(), TR_Memory::ROMClass));
-   if (!romClass)
-      throw std::bad_alloc();
-   memcpy(romClass, &romClassStr[0], romClassStr.size());
-   return romClass;
-   }
-
-J9ROMClass *
-JITServerHelpers::getRemoteROMClass(J9Class *clazz, JITServer::ServerStream *stream, TR_Memory *trMemory, ClassInfoTuple *classInfoTuple)
-   {
-   stream->write(JITServer::MessageType::ResolvedMethod_getRemoteROMClassAndMethods, clazz);
-   const auto &recv = stream->read<ClassInfoTuple>();
-   *classInfoTuple = std::get<0>(recv);
-   return romClassFromString(std::get<0>(*classInfoTuple), trMemory->trPersistentMemory());
-   }
-
-// Return true if able to get data from cache, return false otherwise
-bool
-JITServerHelpers::getAndCacheRAMClassInfo(J9Class *clazz, ClientSessionData *clientSessionData, JITServer::ServerStream *stream, ClassInfoDataType dataType, void *data)
-   {
-   JITServerHelpers::ClassInfoTuple classInfoTuple;
-   ClientSessionData::ClassInfo classInfo;
-   if (!clazz)
-      {
-      return false;
-      }
-      {
-      OMR::CriticalSection getRemoteROMClass(clientSessionData->getROMMapMonitor());
-      auto it = clientSessionData->getROMClassMap().find((J9Class*)clazz);
-      if (it != clientSessionData->getROMClassMap().end())
-         {
-         JITServerHelpers::getROMClassData(it->second, dataType, data);
-         return true;
-         }
-      }
-   stream->write(JITServer::MessageType::ResolvedMethod_getRemoteROMClassAndMethods, clazz);
-   const auto &recv = stream->read<ClassInfoTuple>();
-   classInfoTuple = std::get<0>(recv);
-
-   OMR::CriticalSection cacheRemoteROMClass(clientSessionData->getROMMapMonitor());
-   auto it = clientSessionData->getROMClassMap().find(clazz);
-   if (it == clientSessionData->getROMClassMap().end())
-      {
-      auto romClass = romClassFromString(std::get<0>(classInfoTuple), TR::comp()->trMemory()->trPersistentMemory());
-      JITServerHelpers::cacheRemoteROMClass(clientSessionData, clazz, romClass, &classInfoTuple, classInfo);
-      JITServerHelpers::getROMClassData(classInfo, dataType, data);
-      }
-   else
-      {
-      JITServerHelpers::getROMClassData(it->second, dataType, data);
-      }
-   return false;
-   }
-
-// Return true if able to get data from cache, return false otherwise
-bool
-JITServerHelpers::getAndCacheRAMClassInfo(J9Class *clazz, ClientSessionData *clientSessionData, JITServer::ServerStream *stream, ClassInfoDataType dataType1, void *data1, ClassInfoDataType dataType2, void *data2)
-   {
-   JITServerHelpers::ClassInfoTuple classInfoTuple;
-   ClientSessionData::ClassInfo classInfo;
-   if (!clazz)
-      {
-      return false;
-      }
-      {
-      OMR::CriticalSection getRemoteROMClass(clientSessionData->getROMMapMonitor());
-      auto it = clientSessionData->getROMClassMap().find((J9Class*)clazz);
-      if (it != clientSessionData->getROMClassMap().end())
-         {
-         JITServerHelpers::getROMClassData(it->second, dataType1, data1);
-         JITServerHelpers::getROMClassData(it->second, dataType2, data2);
-         return true;
-         }
-      }
-   stream->write(JITServer::MessageType::ResolvedMethod_getRemoteROMClassAndMethods, clazz);
-   const auto &recv = stream->read<ClassInfoTuple>();
-   classInfoTuple = std::get<0>(recv);
-
-   OMR::CriticalSection cacheRemoteROMClass(clientSessionData->getROMMapMonitor());
-   auto it = clientSessionData->getROMClassMap().find(clazz);
-   if (it == clientSessionData->getROMClassMap().end())
-      {
-      auto romClass = romClassFromString(std::get<0>(classInfoTuple), TR::comp()->trMemory()->trPersistentMemory());
-      JITServerHelpers::cacheRemoteROMClass(clientSessionData, clazz, romClass, &classInfoTuple, classInfo);
-      JITServerHelpers::getROMClassData(classInfo, dataType1, data1);
-      JITServerHelpers::getROMClassData(classInfo, dataType2, data2);
-      }
-   else
-      {
-      JITServerHelpers::getROMClassData(it->second, dataType1, data1);
-      JITServerHelpers::getROMClassData(it->second, dataType2, data2);
-      }
-   return false;
-   }
-
-void
-JITServerHelpers::getROMClassData(const ClientSessionData::ClassInfo &classInfo, ClassInfoDataType dataType, void *data)
-   {
-   switch (dataType)
-      {
-      case CLASSINFO_ROMCLASS_MODIFIERS :
-         {
-         *(uint32_t *)data = classInfo._romClass->modifiers;
-         }
-         break;
-      case CLASSINFO_ROMCLASS_EXTRAMODIFIERS :
-         {
-         *(uint32_t *)data = classInfo._romClass->extraModifiers;
-         }
-         break;
-      case CLASSINFO_BASE_COMPONENT_CLASS :
-         {
-         *(TR_OpaqueClassBlock **)data = classInfo._baseComponentClass;
-         }
-         break;
-      case CLASSINFO_NUMBER_DIMENSIONS :
-         {
-         *(int32_t *)data = classInfo._numDimensions;
-         }
-         break;
-      case CLASSINFO_PARENT_CLASS :
-         {
-         *(TR_OpaqueClassBlock **)data = classInfo._parentClass;
-         }
-         break;
-      case CLASSINFO_CLASS_HAS_FINAL_FIELDS :
-         {
-         *(bool *)data = classInfo._classHasFinalFields;
-         }
-         break;
-      case CLASSINFO_CLASS_DEPTH_AND_FLAGS :
-         {
-         *(uintptrj_t *)data = classInfo._classDepthAndFlags;
-         }
-         break;
-      case CLASSINFO_CLASS_INITIALIZED :
-         {
-         *(bool *)data = classInfo._classInitialized;
-         }
-         break;
-      case CLASSINFO_BYTE_OFFSET_TO_LOCKWORD :
-         {
-         *(uint32_t *)data = classInfo._byteOffsetToLockword;
-         }
-         break;
-      case CLASSINFO_LEAF_COMPONENT_CLASS :
-         {
-         *(TR_OpaqueClassBlock **)data = classInfo._leafComponentClass;
-         }
-         break;
-      case CLASSINFO_CLASS_LOADER :
-         {
-         *(void **)data = classInfo._classLoader;
-         }
-         break;
-      case CLASSINFO_HOST_CLASS :
-         {
-         *(TR_OpaqueClassBlock **)data = classInfo._hostClass;
-         }
-         break;
-      case CLASSINFO_COMPONENT_CLASS :
-         {
-         *(TR_OpaqueClassBlock **)data = classInfo._componentClass;
-         }
-         break;
-      case CLASSINFO_ARRAY_CLASS :
-         {
-         *(TR_OpaqueClassBlock **)data = classInfo._arrayClass;
-         }
-         break;
-      case CLASSINFO_TOTAL_INSTANCE_SIZE :
-         {
-         *(uintptrj_t *)data = classInfo._totalInstanceSize;
-         }
-         break;
-      case CLASSINFO_REMOTE_ROM_CLASS :
-         {
-         *(J9ROMClass **)data = classInfo._remoteRomClass;
-         }
-         break;
-      case CLASSINFO_CLASS_FLAGS :
-         {
-         *(uintptrj_t *)data = classInfo._classFlags;
-         }
-         break;
-      case CLASSINFO_METHODS_OF_CLASS :
-         {
-         *(J9Method **)data = classInfo._methodsOfClass;
-         }
-         break;
-      default :
-         {
-         TR_ASSERT(0, "Class Info not supported %u \n", dataType);
-         }
-         break;
-      }
-   }
-
-void
-JITServerHelpers::postStreamFailure(OMRPortLibrary *portLibrary)
-   {
-   OMR::CriticalSection handleStreamFailure(getClientStreamMonitor());
-
-   OMRPORT_ACCESS_FROM_OMRPORT(portLibrary);
-   uint64_t current_time = omrtime_current_time_millis();
-   if (current_time >= _nextConnectionRetryTime)
-      {
-      _waitTime *= 2; // exponential backoff
-      }
-   _nextConnectionRetryTime = current_time + _waitTime;
-   _serverAvailable = false;
-   }
-
-void
-JITServerHelpers::postStreamConnectionSuccess()
-   {
-   _serverAvailable = true;
-   _waitTime = 1000; // ms
-   }
-
-bool
-JITServerHelpers::shouldRetryConnection(OMRPortLibrary *portLibrary)
-   {
-   OMRPORT_ACCESS_FROM_OMRPORT(portLibrary);
-   return omrtime_current_time_millis() > _nextConnectionRetryTime;
-   }
-
-J9ROMMethod *
-JITServerHelpers::romMethodOfRamMethod(J9Method* method)
-   {
-   if (!TR::CompilationInfo::getStream()) // Not JITServer
-      return J9_ROM_METHOD_FROM_RAM_METHOD((J9Method *)method);
-
-   // JITServer
-   auto clientData = TR::compInfoPT->getClientData();
-   J9ROMMethod *romMethod = NULL;
-
-   // check if the method is already cached
-      {
-      OMR::CriticalSection romCache(clientData->getROMMapMonitor());
-      auto &map = clientData->getJ9MethodMap();
-      auto it = map.find((J9Method*) method);
-      if (it != map.end())
-         romMethod = it->second._romMethod;
-      }
-
-   // if not, go cache the associated ROM class and get the ROM method from it
-   if (!romMethod)
-      {
-      auto stream = TR::CompilationInfo::getStream();
-      stream->write(JITServer::MessageType::VM_getClassOfMethod, (TR_OpaqueMethodBlock*) method);
-      J9Class *clazz = (J9Class*) std::get<0>(stream->read<TR_OpaqueClassBlock *>());
-      TR::compInfoPT->getAndCacheRemoteROMClass(clazz);
-         {
-         OMR::CriticalSection romCache(clientData->getROMMapMonitor());
-         auto &map = clientData->getJ9MethodMap();
-         auto it = map.find((J9Method *) method);
-         if (it != map.end())
-            romMethod = it->second._romMethod;
-         }
-      }
-   TR_ASSERT(romMethod, "Should have acquired romMethod");
-   return romMethod;
-   }
-
 TR::CompilationInfoPerThreadRemote::CompilationInfoPerThreadRemote(TR::CompilationInfo &compInfo, J9JITConfig *jitConfig, int32_t id, bool isDiagnosticThread)
    : CompilationInfoPerThread(compInfo, jitConfig, id, isDiagnosticThread),
    _recompilationMethodInfo(NULL),
@@ -3600,10 +3064,10 @@ TR::CompilationInfoPerThreadRemote::waitForMyTurn(ClientSessionData *clientSessi
          if (TR::Options::isAnyVerboseOptionSet(TR_VerboseCompFailure, TR_VerboseJITServer, TR_VerbosePerformance))
             TR_VerboseLog::writeLineLocked(TR_Vlog_JITServer, "compThreadID=%d timed-out while waiting for seqNo=%d (entry=%p)",
                getCompThreadId(), clientSession->getExpectedSeqNo(), &entry);
-        
+
          // The simplest thing is to delete the cached data for the session and start fresh
          // However, we must wait for any active threads to drain out
-         clientSession->getSequencingMonitor()->enter(); 
+         clientSession->getSequencingMonitor()->enter();
 
          // This thread could have been waiting, then timed-out and then waited
          // to acquire the sequencing monitor. Check whether it can proceed.
@@ -3777,7 +3241,7 @@ TR::CompilationInfoPerThreadRemote::processEntry(TR_MethodToBeCompiled &entry, J
          {
          // Note that it is possible for seqNo to be smaller than expectedSeqNo.
          // This could happen for instance for the very first message that arrives late.
-         // In that case, the second message would have created the client session and 
+         // In that case, the second message would have created the client session and
          // written its own seqNo into expectedSeqNo. We should avoid processing stale updates
          if (TR::Options::isAnyVerboseOptionSet(TR_VerboseCompFailure, TR_VerboseJITServer, TR_VerbosePerformance))
             TR_VerboseLog::writeLineLocked(TR_Vlog_JITServer, "Discarding older msg for clientUID=%llu seqNo=%u < expectedSeqNo=%u compThreadID=%d",
@@ -3785,7 +3249,7 @@ TR::CompilationInfoPerThreadRemote::processEntry(TR_MethodToBeCompiled &entry, J
          clientSession->getSequencingMonitor()->exit();
          throw JITServer::StreamOOO();
          }
-      // We can release the sequencing monitor now because nobody with a 
+      // We can release the sequencing monitor now because nobody with a
       // larger sequence number can pass until I increment expectedSeqNo
       clientSession->getSequencingMonitor()->exit();
 
@@ -4063,7 +3527,7 @@ TR::CompilationInfoPerThreadRemote::processEntry(TR_MethodToBeCompiled &entry, J
    // Decrement number of active threads before _inUse, but we
    // need to acquire the sequencing monitor when accessing numActiveThreads
    getClientData()->getSequencingMonitor()->enter();
-   getClientData()->decNumActiveThreads(); 
+   getClientData()->decNumActiveThreads();
    getClientData()->getSequencingMonitor()->exit();
    getClientData()->decInUse();  // We have the compMonitor so it's safe to access the inUse counter
    if (getClientData()->getInUse() == 0)
@@ -4143,7 +3607,6 @@ TR::CompilationInfoPerThreadRemote::processEntry(TR_MethodToBeCompiled &entry, J
          compInfo->setSuspendThreadDueToLowPhysicalMemory(false);
       }
 
-   
    if (enableJITServerPerCompConn)
       {
       // Delete server stream
@@ -4170,7 +3633,7 @@ TR::CompilationInfoPerThreadRemote::cacheIProfilerInfo(TR_OpaqueMethodBlock *met
          // if entry is NULL, create entryMap, but do not cache anything
          initializePerCompilationCache(entryMap);
          }
-      cacheToPerCompilationMap(_methodIPDataPerComp, (J9Method *) method, entryMap); 
+      cacheToPerCompilationMap(_methodIPDataPerComp, (J9Method *) method, entryMap);
       }
    else if (entry)
       {
@@ -4187,7 +3650,7 @@ TR::CompilationInfoPerThreadRemote::getCachedIProfilerInfo(TR_OpaqueMethodBlock 
 
    IPTableHeapEntry *entryMap = NULL;
    TR_IPBytecodeHashTableEntry *ipEntry = NULL;
-   
+
    getCachedValueFromPerCompilationMap(_methodIPDataPerComp, (J9Method *) method, entryMap);
    // methodInfoPresent=true means that we cached info for this method (i.e. entryMap is not NULL),
    // but entry for this bytecode might not exist, which means it's NULL
@@ -4212,7 +3675,7 @@ TR::CompilationInfoPerThreadRemote::cacheResolvedMethod(TR_ResolvedMethodKey key
 // retrieve a resolved method from the cache
 // returns true if the method is cached, sets resolvedMethod and unresolvedInCP to the cached values.
 // returns false otherwise.
-bool 
+bool
 TR::CompilationInfoPerThreadRemote::getCachedResolvedMethod(TR_ResolvedMethodKey key, TR_ResolvedJ9JITServerMethod *owningMethod, TR_ResolvedMethod **resolvedMethod, bool *unresolvedInCP)
    {
    TR_ResolvedMethodCacheEntry methodCacheEntry;
@@ -4260,7 +3723,7 @@ TR::CompilationInfoPerThreadRemote::getCachedResolvedMethod(TR_ResolvedMethodKey
          else
             *resolvedMethod = method ? new (comp->trHeapMemory()) TR_ResolvedJ9JITServerMethod(method, _vm, comp->trMemory(), methodInfo, owningMethod) : 0;
          }
-         
+
       if (*resolvedMethod)
          {
          if (unresolvedInCP)
