+++ runtime/compiler/control/JITServerCompilationThread.hpp
@@ -23,17 +23,13 @@
 #ifndef JITSERVER_COMPILATION_THREAD_H
 #define JITSERVER_COMPILATION_THREAD_H
 
-#include <unordered_map>
 #include "control/CompilationThread.hpp"
-#include "net/ClientStream.hpp"
-#include "env/PersistentCollections.hpp"
+#include "control/JITServerHelpers.hpp"
 #include "env/j9methodServer.hpp"
-#include "env/J9CPU.hpp"
+#include "net/ClientStream.hpp"
 #include "runtime/JITClientSession.hpp"
 
-class TR_PersistentClassInfo;
 class TR_IPBytecodeHashTableEntry;
-struct TR_RemoteROMStringKey;
 
 using IPTableHeapEntry = UnorderedMap<uint32_t, TR_IPBytecodeHashTableEntry*>;
 using IPTableHeap_t = UnorderedMap<J9Method *, IPTableHeapEntry *>;
@@ -41,16 +37,11 @@ using ResolvedMirrorMethodsPersistIP_t = Vector<TR_ResolvedJ9Method *>;
 using ClassOfStatic_t = UnorderedMap<std::pair<TR_OpaqueClassBlock *, int32_t>, TR_OpaqueClassBlock *>;
 using FieldOrStaticAttrTable_t = UnorderedMap<std::pair<TR_OpaqueClassBlock *, int32_t>, TR_J9MethodFieldAttributes>;
 
-size_t methodStringLength(J9ROMMethod *);
-std::string packROMClass(J9ROMClass *, TR_Memory *);
 TR_MethodMetaData *remoteCompile(J9VMThread *, TR::Compilation *, TR_ResolvedMethod *,
    J9Method *, TR::IlGeneratorMethodDetails &, TR::CompilationInfoPerThreadBase *);
 TR_MethodMetaData *remoteCompilationEnd(J9VMThread * vmThread, TR::Compilation *comp, TR_ResolvedMethod * compilee, J9Method * method,
    TR::CompilationInfoPerThreadBase *compInfoPT, const std::string& codeCacheStr, const std::string& dataCacheStr);
 void outOfProcessCompilationEnd(TR_MethodToBeCompiled *entry, TR::Compilation *comp);
-void printJITServerMsgStats(J9JITConfig *);
-void printJITServerCHTableStats(J9JITConfig *, TR::CompilationInfo *);
-void printJITServerCacheStats(J9JITConfig *, TR::CompilationInfo *);
 
 namespace TR
 {
@@ -157,78 +148,5 @@ class CompilationInfoPerThreadRemote : public TR::CompilationInfoPerThread
    UnorderedMap<std::pair<TR_OpaqueClassBlock *, int32_t>, TR_IsUnresolvedString> *_isUnresolvedStrCache;
    }; // class CompilationInfoPerThreadRemote
 } // namespace TR
-
-class JITServerHelpers
-   {
-   public:
-   enum ClassInfoDataType
-      {
-      CLASSINFO_ROMCLASS_MODIFIERS,
-      CLASSINFO_ROMCLASS_EXTRAMODIFIERS,
-      CLASSINFO_BASE_COMPONENT_CLASS,
-      CLASSINFO_NUMBER_DIMENSIONS,
-      CLASSINFO_PARENT_CLASS,
-      CLASSINFO_INTERFACE_CLASS,
-      CLASSINFO_CLASS_HAS_FINAL_FIELDS,
-      CLASSINFO_CLASS_DEPTH_AND_FLAGS,
-      CLASSINFO_CLASS_INITIALIZED,
-      CLASSINFO_BYTE_OFFSET_TO_LOCKWORD,
-      CLASSINFO_LEAF_COMPONENT_CLASS,
-      CLASSINFO_CLASS_LOADER,
-      CLASSINFO_HOST_CLASS,
-      CLASSINFO_COMPONENT_CLASS,
-      CLASSINFO_ARRAY_CLASS,
-      CLASSINFO_TOTAL_INSTANCE_SIZE,
-      CLASSINFO_CLASS_OF_STATIC_CACHE,
-      CLASSINFO_REMOTE_ROM_CLASS,
-      CLASSINFO_CLASS_FLAGS,
-      CLASSINFO_METHODS_OF_CLASS,
-      };
-   // NOTE: when adding new elements to this tuple, add them to the end,
-   // to not mess with the established order.
-   using ClassInfoTuple = std::tuple
-      <
-      std::string, J9Method *,                                       // 0: string                   1: _methodsOfClass
-      TR_OpaqueClassBlock *, int32_t,                                // 2: _baseComponentClass      3: _numDimensions
-      TR_OpaqueClassBlock *, std::vector<TR_OpaqueClassBlock *>,     // 4: _parentClass             5: _tmpInterfaces
-      std::vector<uint8_t>, bool,                                    // 6: _methodTracingInfo       7: _classHasFinalFields
-      uintptrj_t, bool,                                              // 8: _classDepthAndFlags      9: _classInitialized
-      uint32_t, TR_OpaqueClassBlock *,                               // 10: _byteOffsetToLockword   11: _leafComponentClass
-      void *, TR_OpaqueClassBlock *,                                 // 12: _classLoader            13: _hostClass
-      TR_OpaqueClassBlock *, TR_OpaqueClassBlock *,                  // 14: _componentClass         15: _arrayClass
-      uintptrj_t, J9ROMClass *,                                      // 16: _totalInstanceSize      17: _remoteRomClass
-      uintptrj_t, uintptrj_t                                         // 18: _constantPool           19: _classFlags
-      >;
-   static ClassInfoTuple packRemoteROMClassInfo(J9Class *clazz, J9VMThread *vmThread, TR_Memory *trMemory);
-   static void cacheRemoteROMClass(ClientSessionData *clientSessionData, J9Class *clazz, J9ROMClass *romClass, ClassInfoTuple *classInfoTuple);
-   static void cacheRemoteROMClass(ClientSessionData *clientSessionData, J9Class *clazz, J9ROMClass *romClass, ClassInfoTuple *classInfoTuple, ClientSessionData::ClassInfo &classInfo);
-   static J9ROMClass *getRemoteROMClassIfCached(ClientSessionData *clientSessionData, J9Class *clazz);
-   static J9ROMClass *getRemoteROMClass(J9Class *, JITServer::ServerStream *stream, TR_Memory *trMemory, ClassInfoTuple *classInfoTuple);
-   static J9ROMClass *romClassFromString(const std::string &romClassStr, TR_PersistentMemory *trMemory);
-   static bool getAndCacheRAMClassInfo(J9Class *clazz, ClientSessionData *clientSessionData, JITServer::ServerStream *stream, ClassInfoDataType dataType, void *data);
-   static bool getAndCacheRAMClassInfo(J9Class *clazz, ClientSessionData *clientSessionData, JITServer::ServerStream *stream, ClassInfoDataType dataType1, void *data1, ClassInfoDataType dataType2, void *data2);
-   static void getROMClassData(const ClientSessionData::ClassInfo &classInfo, ClassInfoDataType dataType, void *data);
-   static J9ROMMethod *romMethodOfRamMethod(J9Method* method);
-
-   // functions used for allowing the client to compile locally when server is unavailable
-   // should be only used on the client side
-   static void postStreamFailure(OMRPortLibrary *portLibrary);
-   static bool shouldRetryConnection(OMRPortLibrary *portLibrary);
-   static void postStreamConnectionSuccess();
-   static bool isServerAvailable() { return _serverAvailable; }
-   static TR::Monitor * getClientStreamMonitor()
-      {
-      if (!_clientStreamMonitor)
-         _clientStreamMonitor = TR::Monitor::create("clientStreamMonitor");
-      return _clientStreamMonitor;
-      }
-   static void insertIntoOOSequenceEntryList(ClientSessionData *clientData, TR_MethodToBeCompiled *entry);
-
-   private:
-   static uint64_t _waitTime;
-   static uint64_t _nextConnectionRetryTime;
-   static bool _serverAvailable;
-   static TR::Monitor * _clientStreamMonitor;
-   }; // class JITServerHelpers
 
 #endif // defined(JITSERVER_COMPILATION_THREAD_H)
